requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.1
itsdangerous==1.1.0
jinja2>=2.11.3
joblib==0.17.0
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.3
MarkupSafe==1.1.1
numba==0.52.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.7
pip>=21.1
pooch==1.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
requests==2.25.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa>=4.7
scikit-learn==0.23.2
scipy==1.5.4
setuptools==50.3.2
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu>=2.5.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3>=1.26.5
Werkzeug==1.0.1
wheel==0.36.0
wrapt==1.12.1
zipp==3.4.0
